Reimplement test and improve API
