+++ dsh_sdk/src/certificates/mod.rs
@@ -279,15 +279,6 @@ impl Cert {
     }
 }
 
-/// Helper function to ensure that the host starts with `https://` or `http://`.
-pub(crate) fn ensure_https_prefix(host: impl AsRef<str>) -> String {
-    if host.as_ref().starts_with("http://") || host.as_ref().starts_with("https://") {
-        host.as_ref().to_string()
-    } else {
-        format!("https://{}", host.as_ref())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -384,20 +375,5 @@ mod tests {
             cert.private_key_pem().as_bytes(),
         );
         assert!(identity.is_ok());
-    }
-
-    #[test]
-    fn test_ensure_https_prefix() {
-        let host = "http://example.com";
-        let result = ensure_https_prefix(host);
-        assert_eq!(result, "http://example.com");
-
-        let host = "https://example.com";
-        let result = ensure_https_prefix(host);
-        assert_eq!(result, "https://example.com");
-
-        let host = "example.com";
-        let result = ensure_https_prefix(host);
-        assert_eq!(result, "https://example.com");
     }
 }
