+++ dsh_sdk/src/dsh.rs
@@ -28,7 +28,7 @@ use log::warn;
 use std::env;
 use std::sync::{Arc, OnceLock};
 
-use crate::certificates::{ensure_https_prefix, Cert, CertificatesError};
+use crate::certificates::{Cert, CertificatesError};
 use crate::datastream::Datastream;
 use crate::error::DshError;
 use crate::utils;
@@ -113,7 +113,7 @@ impl Dsh {
         let tenant_name = utils::tenant_name().unwrap_or_else(|_| "local_tenant".to_string());
         let task_id =
             utils::get_env_var(VAR_TASK_ID).unwrap_or_else(|_| "local_task_id".to_string());
-        let config_host = utils::get_env_var(VAR_KAFKA_CONFIG_HOST).map(ensure_https_prefix);
+        let config_host = utils::get_env_var(VAR_KAFKA_CONFIG_HOST).map(utils::ensure_https_prefix);
 
         let certificates = if let Ok(cert) = Cert::from_pki_config_dir::<std::path::PathBuf>(None) {
             Some(cert)
