+++ dsh_sdk/src/protocol_adapters/token/error.rs
